Update README.md
The hackathon theme Localvore.Guide applies to is "Problems facing our planet" because we feel keeping a thriving local economy is key to sustaining the world as software eats the world and businesses become more efficient. We need to help local businesses and entrepreneurs thrive. Localvore.Guide does this by presenting local businesses in a manner that makes them easy to find and engage with.